% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStartEndDate.r
\name{checkStartEndDate}
\alias{checkStartEndDate}
\title{checkStartEndDate}
\usage{
checkStartEndDate(startDate, endDate, interactive = TRUE)
}
\arguments{
\item{startDate}{character}

\item{endDate}{character}

\item{interactive}{logical Option for interactive mode.  If true, there is user interaction for error handling and data checks.}
}
\value{
vector where first value is startDate, second is endDate
}
\description{
Checks that the start date is before the end date.  If not, it will give the user the opportunity to correct, otherwise will create a warning.
}
\keyword{WRTDS}
\keyword{flow}
