% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSeries.R
\name{checkSurfaceSpan}
\alias{checkSurfaceSpan}
\title{checkSurfaceSpan}
\usage{
checkSurfaceSpan(eList)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}
}
\description{
checkSurfaceSpan
}
