\docType{data}
\name{ChopDaily}
\alias{ChopDaily}
\title{Example Daily dataframe for Choptank River}
\description{
  Example data representing streamflowfrom the Choptank
  River at Greensboro, MD, USGS data as agumented by the
  modelEstimation process.
}
\author{
  Robert M. Hirsch \email{rhirsch@usgs.gov}
}
\keyword{data}
\keyword{datasets}
\keyword{quality}
\keyword{streamflow}
\keyword{water}

