\docType{data}
\name{ChopSample}
\alias{ChopSample}
\title{Example Sample dataframe for Choptank River}
\description{
  Example data representing Nitrate sample data from the
  Choptank River at Greensboro, MD, USGS data as agumented
  by the modelEstimation process
}
\author{
  Robert M. Hirsch \email{rhirsch@usgs.gov}
}
\keyword{data}
\keyword{datasets}
\keyword{quality}
\keyword{water}

