% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Choptank_eList}
\alias{Choptank_eList}
\title{Example eList}
\description{
Example data representing data from the Choptank River at Greensboro, MD,  USGS data
Data is a named list of the Daily, Sample, INFO dataframes, and the surface matrix.
}
\keyword{data}
\keyword{quality}
\keyword{water}

