% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxUnit.R
\docType{class}
\name{concUnit-class}
\alias{concUnit-class}
\title{concUnit class}
\description{
Some details about the concUnit class
}
\details{
\describe{
   \item{longPrefix}{A character specifying the long name
    for concentration labels.}

   \item{shortPrefix}{A character specifying the short name
    for concentration labels.}

 }
}
