% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGRET.R
\docType{data}
\name{Constants}
\alias{Constants}
\alias{fluxConst}
\alias{qConst}
\alias{monthInfo}
\alias{concConst}
\title{Constants included with EGRET}
\description{
\tabular{ll}{
fluxConst \tab Flux conversion object\cr
qConst \tab Flow conversion object\cr
monthInfo \tab Month object\cr
}
}
\examples{
fluxConst
fluxConst[['kgDay']]
fluxConst[['kgDay']]@unitName
qConst
qConst[['cfs']]
qConst[['cfs']]@qUnitName
concConst[['concentration']]
concConst[['concentration']]@shortPrefix
}
\keyword{datasets}
