% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumQdate.R
\name{cumQdate}
\alias{cumQdate}
\title{Cumulative flow calculation}
\usage{
cumQdate(eList, paStart = 10, paLong = 12, fract = 0.5)
}
\arguments{
\item{eList}{named list with at least the Sample and INFO dataframes}

\item{paStart}{numeric integer specifying the starting month for the period of analysis, 1 <= paStart <= 12, default is 10}

\item{paLong}{numeric integer specifying the length of the period of analysis, in months, 1 <= paLong <= 12, default is 12}

\item{fract}{numeric fraction of the flow}
}
\description{
This function computes the first day of the calendar year
at which a specific fraction of the cumulative flow for that year
has been exceeded.  Typically one looks for the point where
half the cumulative flow has happened (fract = 0.5)
the portion of the year being considered are set by paStart and paLong
the matrix returned has 2 columns: 
the first is the year (integer when the period of analysis ends),
the second is the day of the year when the fraction has been exceeded.
None of the rows will have any NA values.
}
\examples{
eList <- Choptank_eList
annualFlow <- cumQdate(eList)
head(annualFlow)
plot(annualFlow)
mod1 <- lm(annualFlow[,2] ~ annualFlow[,1])
summary(mod1)
}
