% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{decimalHighLow}
\alias{decimalHighLow}
\title{decimalHighLow}
\usage{
decimalHighLow(df)
}
\arguments{
\item{df}{data.frame with Month, DecYear, and Month columns}
}
\value{
list with DecHigh and DecLow (water year high/low decimal values)
}
\description{
decimalHighLow figures out the highest and lowest decimal year based on
water year. The input is a data frame with columns Month and DecYear.
}
\examples{
eList <- Choptank_eList
highLow <- decimalHighLow(eList$Sample)

DecHigh <- highLow[["DecHigh"]]
DecLow <- highLow[["DecLow"]]
}
