\docType{package}
\name{EGRET-package}
\alias{EGRET-package}
\title{EGRET package (except for dataRetrieval) includes WRTDS and flowHistory}
\description{
  \tabular{ll}{ Package: \tab EGRET\cr Type: \tab
  Package\cr Version: \tab 0.1\cr Date: \tab 2012-04-23\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr }
}
\details{
  Collection of functions to do WRTDS and flowHistory
  analysis, and produce graphs and tables of data and
  results from these analyses.
}
\author{
  Robert M. Hirsch \email{rhirsch@usgs.gov}, Laura De Cicco
  \email{ldecicco@usgs.gov}
}
\references{
  Hirsch, R. M., Moyer, D. L. and Archfield, S. A. (2010),
  Weighted Regressions on Time, Discharge, and Season
  (WRTDS), with an Application to Chesapeake Bay River
  Inputs. JAWRA Journal of the American Water Resources
  Association, 46: 857-880. doi:
  10.1111/j.1752-1688.2010.00482.x
}
\keyword{graphics}
\keyword{statistics}
\keyword{streamflow}
\keyword{water-quality}

