% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorStats.R
\name{errorStats}
\alias{errorStats}
\title{Error statistics}
\usage{
errorStats(eList)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}
}
\description{
This function takes a fitted WRTDS model and computes error statistics
the residuals used here are cross-validation residuals, 
which will be slightly larger than regular regression residuals
in the case of censored data, the residuals are computed from random residuals
computed from makeAugmentedSample()
}
\examples{
eList <- Choptank_eList
errorStats(eList)
}
