% Generated by roxygen2 (4.0.2): do not edit by hand
\name{estDailyWithoutNormalization}
\alias{estDailyWithoutNormalization}
\title{Estimates all daily values of Concentration, and Flux}
\usage{
estDailyWithoutNormalization(eList)
}
\arguments{
\item{eList}{named list with at least Daily and INFO dataframes, and surface matrix.}
}
\value{
localDaily string specifying the name of the data frame containing the daily values and these esimates
}
\description{
Uses the surfaces estimated in estSurfaces to estimate these two time series
   in addition to the time series for standard error and yHat (estimated log concentration).
   The results are stored in an augemented version of the Daily data frame, which is returned.
   This code is identical to estDailyFromSurfaces but it lacks the flow normalization process.
   The exclusion of the flow-normalization process saves a large amount of computer time.
}
\examples{
eList <- Choptank_eList
Daily <- estDailyWithoutNormalization(eList)
}
\keyword{statistics}
\keyword{water-quality}

