% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexFNnew.R
\name{estFNyears}
\alias{estFNyears}
\title{Estimates for flow normalization by year}
\usage{
estFNyears(eList, years, waterYear = TRUE)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{years}{vector of years}
}
\description{
Estimates for flow normalization by year
}
