\docType{data}
\name{exAnnualResults}
\alias{exAnnualResults}
\title{Example AnnualResults dataframe}
\description{
  Example data representing annual WRTDS results for
  Nitrate data from the Choptank River at Greensboro, MD,
  USGS data as output from the function setupYears
}
\author{
  Robert M. Hirsch \email{rhirsch@usgs.gov}
}
\keyword{data}
\keyword{datasets}
\keyword{quality}
\keyword{water}

