\docType{data}
\name{exannualSeries}
\alias{exannualSeries}
\title{Example annualSeries dataframe}
\description{
  Example data representing annual Series of streamflow
  statistics from the Choptank River at Greensboro, MD,
  USGS data as output from the function makeAnnualSeries
}
\author{
  Robert M. Hirsch \email{rhirsch@usgs.gov}
}
\keyword{data}
\keyword{datasets}
\keyword{streamflow}

