\docType{data}
\name{exDailyEnd}
\alias{exDailyEnd}
\title{Data included in EGRET}
\description{
  Example data representing Streamflow and Nitrate from the
  Choptank River at Greensboro, MD, USGS data as agumented
  by the modelEstimation process
}
\author{
  Robert M. Hirsch \email{rhirsch@usgs.gov}
}
\keyword{data}
\keyword{datasets}
\keyword{quality}
\keyword{streamflow}
\keyword{water}

