\docType{data}
\name{exDailyStart}
\alias{exDailyStart}
\title{Example initial Daily dataframe}
\description{
  Example data representing streamflow from the Choptank
  River at Greensboro, MD, USGS data. This is the data that
  is initially generated from the dataRetrieval package.
}
\author{
  Robert M. Hirsch \email{rhirsch@usgs.gov}
}
\keyword{data}
\keyword{datasets}
\keyword{quality}
\keyword{streamflow}
\keyword{water}

