\docType{data}
\name{exINFOStart}
\alias{exINFOStart}
\title{Data included in EGRET}
\description{
  Example data representing meta-data from the Choptank
  River at Greensboro, MD, USGS data
}
\author{
  Robert M. Hirsch \email{rhirsch@usgs.gov}
}
\keyword{data}
\keyword{datasets}
\keyword{quality}
\keyword{water}

