\docType{data}
\name{exMonthlyResults}
\alias{exMonthlyResults}
\title{Example MonthlyResults dataframe}
\description{
  Example data representing monthly WRTDS results for
  Nitrate data from the Choptank River at Greensboro, MD,
  USGS data as output from the function
  calculateMonthlyResults
}
\author{
  Robert M. Hirsch \email{rhirsch@usgs.gov}
}
\keyword{data}
\keyword{datasets}
\keyword{quality}
\keyword{water}

