\docType{data}
\name{exSampleStart}
\alias{exSampleStart}
\title{Data included in EGRET}
\description{
  Example data representing Nitrate sample data from the
  Choptank River at Greensboro, MD, USGS data
}
\author{
  Robert M. Hirsch \email{rhirsch@usgs.gov}
}
\keyword{data}
\keyword{datasets}
\keyword{quality}
\keyword{water}

