\docType{data}
\name{exsurfaces}
\alias{exsurfaces}
\title{Data included in EGRET}
\description{
  Example data representing surfaces, from the fitted model
  of Nitrate data for the Choptank River at Greensboro, MD,
  USGS data output from the function estSurfaces
}
\author{
  Robert M. Hirsch \email{rhirsch@usgs.gov}
}
\keyword{data}
\keyword{datasets}
\keyword{quality}
\keyword{water}

