% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillUpElists.R
\name{fillUpElists}
\alias{fillUpElists}
\title{fillUpElists}
\usage{
fillUpElists(eList)
}
\arguments{
\item{eList}{named list with the INFO, Daily, and Sample dataframes and surfaces matrix}
}
\description{
Function used in `runSeries` to populate the yHat, SE, and ConcHat
columns into the Daily and Sample data frames within the eList.
}
\examples{

eList <- Choptank_eList
eList$Daily$yHat <- NULL
eList$Daily$SE <- NULL
eList$Sample$SE <- NULL
eList$Sample$yHat <- NULL

eList <- fillUpElists(eList)
head(eList$Daily$yHat)
head(eList$Daily$SE)
head(eList$Sample$SE)
head(eList$Sample$yHat)

}
