% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexFN.R
\name{flexFN}
\alias{flexFN}
\title{Flexible Flow Normalization}
\usage{
flexFN(eList, dateInfo, sampleStartCol = "sampleStart",
  sampleEndCol = "sampleEnd", flowStartCol = "flowStart",
  flowEndCol = "flowEnd")
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{dateInfo}{data frame with 4 columns, their names defined by sampleStart, sampleEnd, flowStart, flowEnd}

\item{sampleStart}{integer vector of start years (water) for each FN conc/flux segment}

\item{flowStart}{integer vector of start years (water) for flow normalization}

\item{flowEnd}{integer vector of end years (water) for flow normalization}
}
\description{
Flexible Flow Normalization
}
\examples{
eList <- Choptank_eList
eList <- setUpEstimation(eList)
sampleStart <- c("1980-01-01","1990-01-01","2000-01-01")
sampleEnd <- c("1995-06-06","2002-09-30","2011-02-01")
flowStart <- c("1980-02-01","1985-06-01","1992-10-10")
flowEnd <- c("1994-06-15","2004-03-03","2009-03-15")
dateInfo <- data.frame(sampleStart,
                       sampleEnd,
                       flowStart, 
                       flowEnd, stringsAsFactors = FALSE)
\dontrun{
newEList <- flexFN(eList, dateInfo)
plotFluxHist(newEList)
flexPlotAddOn(newEList)
}
}
