\name{fluxUnit-class}
\alias{fluxUnit-class}
\title{fluxUnit class}
\description{
  Some details about the fluxUnit class
}
\details{
  \describe{ \item{shortName}{A character specifying the
  short name.}

  \item{unitFactor}{A numeric representing the conversion
  factor}

  \item{unitName}{A character specifying the full name.}

  \item{unitExpress}{An expression specifying the full name
  starting with Observed.}

  \item{unitEstimate}{An expression specifying the full
  name starting with Estimated.}

  \item{shortCode}{A number for quick lookup} }
}

