% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatCheckDate.r
\name{formatCheckDate}
\alias{formatCheckDate}
\title{formatCheckDate}
\usage{
formatCheckDate(Date, dateString, interactive = TRUE)
}
\arguments{
\item{Date}{character}

\item{dateString}{character used in either error message or interactive message. An example would be "startDate"}

\item{interactive}{logical Option for interactive mode.  If true, there is user interaction for error handling and data checks.}
}
\value{
condition logical if TRUE,
}
\description{
This function was never incorporated into the EGRET workflow and will be removed in future versions.
}
\details{
Response to the date format checker.  If the date is not formated correctly, it will give the user the opportunity to correct, otherwise will create a warning.
}
\keyword{WRTDS}
\keyword{flow}
