% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_functions.R
\name{genmissing}
\alias{genmissing}
\title{genmissing}
\usage{
genmissing(X1, XN, rho, N)
}
\arguments{
\item{X1}{value before the gap}

\item{XN}{value after the gap}

\item{rho}{the lag one autocorrelation}

\item{N}{the length of the sequence including X1 and XN. It
is two more than the gap length}
}
\value{
genmissing numeric vector of length N, conditioned on the
first value (X1) and last value (XN) with the specified lag one autocorrelation
in the limit (where N is large) the values are normal with mean 0 and variance 1
}
\description{
Generates a lag one auto-regressive time series, where the first and last
values are fixed.  Marginal expected value is zero and variance is one.  
Generated values have a normal conditional distribution.
}
\author{
Tim Cohn
}
\keyword{internal}
