% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estDailyFromSurfacesJ.R
\name{getConcFluxFromSurface}
\alias{getConcFluxFromSurface}
\title{getConcFluxFromSurface}
\usage{
getConcFluxFromSurface(eList, allLogQsByDayOfYear, localDaily,
  localsurfaces = NA)
}
\arguments{
\item{Daily}{data frame}
}
\description{
getConcFluxFromSurface
}
