% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeReport.r
\name{getSample}
\alias{getSample}
\alias{getSample.default}
\title{Get Sample dataframe from EGRET object}
\usage{
getSample(x, ...)

getSample(x, ...)

getSample.default(x, ...)
}
\arguments{
\item{x}{EGRET object or named list}

\item{\dots}{additional parameters}
}
\value{
Sample dataframe
}
\description{
From a named list or EGRET object, extract the Sample dataframe
}
\examples{
eList <- Choptank_eList
Sample <- getSample(eList)
}
\seealso{
\code{\link{readNWISDaily}}, \code{\link{readNWISSample}}
}
\keyword{USGS}
\keyword{WRTDS}
\keyword{data}
\keyword{import}

