% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estDailyFromSurfacesJ.R
\name{getSurfaceEstimates}
\alias{getSurfaceEstimates}
\title{getSurfaceEstimates}
\usage{
getSurfaceEstimates(eList, localsurfaces = NA, localDaily = NA)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}
}
\value{
Daily dataframe with yHat, SE, ConcDay and FluxDay calulated
}
\description{
getSurfaceEstimates
}
