% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergeReport.r
\name{getSurfaces}
\alias{getSurfaces}
\alias{getSurfaces.default}
\alias{getSurfaces.egret}
\title{Get surfaces matrix from EGRET object}
\usage{
getSurfaces(x, ...)

\method{getSurfaces}{egret}(x, ...)

\method{getSurfaces}{default}(x, ...)
}
\arguments{
\item{x}{EGRET object or named list}

\item{\dots}{additional parameters}
}
\value{
Sample dataframe
}
\description{
From a named list or EGRET object, extract the surfaces matrix
}
\examples{
eList <- Choptank_eList
surfaces <- getSurfaces(eList)
}
\seealso{
\code{\link{readNWISDaily}}, \code{\link{readNWISSample}}
}
\keyword{USGS}
\keyword{WRTDS}
\keyword{data}
\keyword{import}

