% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egret_objects.R
\name{print.egret}
\alias{print.egret}
\alias{plot.egret}
\alias{nDischarge}
\alias{nObservations}
\alias{nCensoredVals}
\title{EGRET helper functions}
\usage{
\method{print}{egret}(x, ...)

\method{plot}{egret}(x, ...)

nDischarge(x)

nObservations(x)

nCensoredVals(x)
}
\arguments{
\item{x}{EGRET object}

\item{\dots}{additional parameters}
}
\description{
A small collection of helper functions
}
\examples{
Choptank_eList
print(Arkansas_eList)
plot(Choptank_eList)
plot(Choptank_eList, cex.main=0.7)
nDischarge(Arkansas_eList)
nObservations(Arkansas_eList)
nCensoredVals(Arkansas_eList)
}
\seealso{
\code{\link{multiPlotDataOverview}}
}
\keyword{USGS}
\keyword{WRTDS}
\keyword{data}
\keyword{import}
