% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mergeReport.r
\name{is.egret}
\alias{is.egret}
\title{Check for EGRET object}
\usage{
is.egret(x)
}
\arguments{
\item{x}{object to check}
}
\value{
logical
}
\description{
Checks object to see if it is an EGRET object
}
\examples{
eList <- Choptank_eList
is.egret(eList)
}
\keyword{USGS}
\keyword{WRTDS}
\keyword{data}
\keyword{import}

