% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSurvReg.R
\name{jitterSam}
\alias{jitterSam}
\title{jitter Sample}
\usage{
jitterSam(Sam, V = 0.2)
}
\arguments{
\item{Sam}{data frame with at least columns DecYear and LogQ}

\item{V}{a multiplier for the sd of the LogQ jitter. for example V = 0.02,
means that the sd of the LnQ jitter is 0.02*sdLQ}
}
\description{
jitter Sample
}
