% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logPretty1.R
\name{logPretty1}
\alias{logPretty1}
\title{Sets up tick marks for an axis with a log scale, where the graph is small}
\usage{
logPretty1(xMin, xMax)
}
\arguments{
\item{xMin}{A numeric value for the minimum value to be plotted, it must be >0}

\item{xMax}{A numeric value for the maximum value to be plotted, it must be >xMax}
}
\value{
xTicks A vector representing the values for each of the tick marks
}
\description{
Axis tick marks for a log scale for cases where the data cover many orders of magnitude
and the graph is small. 
These tick marks are designed to progress by factors of 10.
}
\examples{
xMin<-0.7
xMax<-990000
logPretty1(xMin,xMax)
xMin<-3
xMax<-15
logPretty1(xMin,xMax)
}
\keyword{graphics}
\keyword{statistics}

