% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logPretty3.R
\name{logPretty3}
\alias{logPretty3}
\title{Sets up tick marks for an axis with a log scale}
\usage{
logPretty3(xMin, xMax)
}
\arguments{
\item{xMin}{A numeric value for the minimum value to be plotted, it must be >0}

\item{xMax}{A numeric value for the maximum value to be plotted, it must be >xMax}
}
\value{
xTicks A vector representing the values for each of the tick marks
}
\description{
Axis tick marks for a log scale.
These tick marks are designed to progress with 3 tick marks for every factor of 10.
For example: 2,5,10,20,50,100,200,500.
}
\examples{
logPretty3(0.7,990000)
logPretty3(3,15)
}
\keyword{graphics}
\keyword{statistics}

