% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSeries.R
\name{makeSeriesOutputs}
\alias{makeSeriesOutputs}
\title{makeSeriesOutputs}
\usage{
makeSeriesOutputs(eList, startDate = NA, endDate = NA, printTable = TRUE,
  plotConc = TRUE, plotFlux = TRUE)
}
\arguments{
\item{eList}{named list with at least Daily and INFO dataframes}

\item{startDate}{character or Date}

\item{endDate}{character or Date}

\item{printTable}{logical}

\item{plotConc}{logical}

\item{plotFlux}{logical}
}
\description{
makeSeriesOutputs
}
\examples{
eList <- Choptank_Phos
\dontrun{
eList_series1 <- runSeries(eList, windowSide = 7)
makeSeriesOutputs(eList_series1)

}
}
