% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mergeReport}
\alias{mergeReport}
\title{Merge Sample and Daily Data for WRTDS}
\usage{
mergeReport(INFO, Daily, Sample, interactive = TRUE)
}
\arguments{
\item{INFO}{dataframe containing the INFO dataframe}

\item{Daily}{dataframe containing the daily data}

\item{Sample}{dataframe containing the sample data}

\item{interactive}{logical Option for interactive mode.  If true, there is user interaction for error handling and data checks.}
}
\value{
newSample dataframe with merged flow information
}
\description{
Merges the flow data from the daily record into the sample record.
}
\examples{
siteNumber <- '01594440'
pCode <- '01075'
Daily <- getNWISDaily(siteNumber,'00060', '1985-01-01', '1985-03-31')
Sample <- getNWISSample(siteNumber,pCode, '1985-01-01', '1985-03-31')
INFO <- getNWISInfo(siteNumber,pCode,interactive=FALSE)
returnedList <- mergeReport(INFO, Daily, Sample)
Sample <- returnedList$Sample
}
\seealso{
\code{\link{getNWISDaily}}, \code{\link{getNWISSample}}
}
\keyword{USGS}
\keyword{WRTDS}
\keyword{data}
\keyword{import}

