\name{monthLabel-class}
\alias{monthLabel-class}
\title{monthLabel class}
\description{
  Some details about the monthLabel class
}
\details{
  \describe{ \item{monthAbbrev}{A character specifying the
  abbreviated month name.}

  \item{monthFull}{A character specifying the full month
  name}

  \item{monthSingle}{A character specifying the single
  letter of the month.} }
}

