% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeReport.r
\name{nCensoredVals}
\alias{nCensoredVals}
\title{Number of censored samples}
\usage{
nCensoredVals(eList)
}
\arguments{
\item{eList}{named list with at least Sample dataframe}
}
\description{
Determine the number of censored sample data points in an eList
}
\examples{
nCensoredVals(Arkansas_eList)
}
