% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeReport.r
\name{nDischarge}
\alias{nDischarge}
\title{Number of discharge values}
\usage{
nDischarge(eList)
}
\arguments{
\item{eList}{named list with at least Daily dataframe}
}
\description{
Determine the number of discharge data points in an eList
}
\examples{
nDischarge(Arkansas_eList)
}
