% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeReport.r
\name{nObservations}
\alias{nObservations}
\title{Number of observations}
\usage{
nObservations(eList)
}
\arguments{
\item{eList}{named list with at least Sample dataframe}
}
\description{
Determine the number of observations from sampled data in an eList
}
\examples{
nObservations(Arkansas_eList)
}
