% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeReport.r
\name{nSamples}
\alias{nSamples}
\title{Number of samples}
\usage{
nSamples(eList)
}
\arguments{
\item{eList}{named list with at least Sample dataframe}
}
\description{
Determine the number of sampled data points in an eList
}
\examples{
nSamples(Arkansas_eList)
}
