% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeReport.r
\name{plot.egret}
\alias{plot.egret}
\title{Plots EGRET object}
\usage{
\method{plot}{egret}(x, ...)
}
\arguments{
\item{x}{named EGRET list with at least Daily, Sample, and INFO dataframes}

\item{\dots}{any other arguments to be passed to \code{multiPlotDataOverview}}
}
\description{
Plot function for EGRET object (calls multiPlotDataOverView).
}
\examples{
eList <- Choptank_eList
plot(eList)
plot(eList, cex.main=0.7)
}
