% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConcFlow.R
\name{plotFlowConc}
\alias{plotFlowConc}
\title{Plot concentration response to flow as a lineplot for all months}
\usage{
plotFlowConc(eList, month = c(1:12), years = NULL, col_vec = c("red",
  "green", "blue"), alpha = 1, size = 1, allflo = FALSE, ncol = NULL,
  grids = TRUE, scales = NULL, pretty = TRUE, use_bw = TRUE,
  fac_nms = NULL)
}
\arguments{
\item{eList}{input egret object}

\item{month}{numeric input from 1 to 12 indicating the monthly predictions to plot}

\item{years}{numeric vector of years to plot, defaults to all}

\item{col_vec}{chr string of plot colors to use, passed to \code{\link[ggplot2]{scale_colour_gradientn}} for line shading.}

\item{alpha}{numeric value from zero to one indicating line transparency}

\item{size}{numeric value for line size}

\item{allflo}{logical indicating if the salinity or flow values for plotting are limited to the fifth and ninety-fifth percentile of observed values for the month of interest}

\item{ncol}{numeric argument passed to \code{\link[ggplot2]{facet_wrap}} indicating number of facet columns}

\item{grids}{logical indicating if grid lines are present}

\item{scales}{chr string passed to ggplot to change x/y axis scaling on facets, acceptable values are \code{'free'}, \code{'free_x'}, or \code{'free_y'}}

\item{pretty}{logical indicating if plot aesthetics are applied, otherwise the \code{\link[ggplot2]{ggplot}} default themes are used.  The aesthetic arguments will not apply if \code{pretty = TRUE}.}

\item{use_bw}{logical indicating if the \code{\link[ggplot2]{theme_bw}} theme is used}

\item{fac_nms}{optional chr string for facet labels, which must be equal in length to \code{month}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that can be further modified
}
\description{
Plot the relationship between the modelled concentration and flow across the time series using line plots for each month.  Each line corresponds to a unique year.  This can be used to evaluate temporal variation between the two.
}
\details{
These plots can be used to examine how the relationship between the response variable and flow varies throughout the time series.  Line plots are returned that show the relationship of the response variable with flow using different lines for each year. The interpolation grid from WRTDS is used to create the plot.  Each plot is limited to the same month throughout the time series to limit seasonal variation.  Plots are also constrained to the fifth and ninety-fifth percentile of observed flow values during the month of interest to limit the predictions within the data domain. This behavior can be suppressed by changing the \code{allflo} argument. 

Note that the year variable used for color mapping is treated as a continuous variable although it is an integer by definition.
}
\examples{

eList <- Choptank_eList

# defaults
plotFlowConc(eList)

# change the defaults
plotFlowConc(eList, month = 2, years = seq(1980, 1990), 
 col_vec = rainbow(7), alpha = 0.5, size = 3) 
}

