% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFluxHist.R
\name{plotFluxHist}
\alias{plotFluxHist}
\title{Graph of annual flux and flow normalized flux versus year}
\usage{
plotFluxHist(eList, yearStart = NA, yearEnd = NA, fluxUnit = 9,
  fluxMax = NA, printTitle = TRUE, usgsStyle = FALSE,
  plotFlowNorm = TRUE, plotAnnual = TRUE, plotGenFlux = FALSE,
  tinyPlot = FALSE, col = "black", col.pred = "green", col.gen = "red",
  cex = 0.8, cex.axis = 1.1, cex.main = 1.1, lwd = 2,
  customPar = FALSE, ...)
}
\arguments{
\item{eList}{named list with at least the Daily and INFO dataframes}

\item{yearStart}{numeric is the calendar year containing the first estimated annual value to be plotted, default is NA (which allows it to be set automatically by the data)}

\item{yearEnd}{numeric is the calendar year just after the last estimated annual value to be plotted, default is NA (which allows it to be set automatically by the data)}

\item{fluxUnit}{number representing entry in pre-defined fluxUnit class array. \code{\link{printFluxUnitCheatSheet}}}

\item{fluxMax}{number specifying the maximum value to be used on the vertical axis, default is NA (which allows it to be set automatically by the data)}

\item{printTitle}{logical variable if TRUE title is printed, if FALSE title is not printed (this is best for a multi-plot figure)}

\item{usgsStyle}{logical option to use USGS style guidelines. Setting this option
to TRUE does NOT guarantee USGS compliance. It will only change automatically
generated labels.}

\item{plotFlowNorm}{logical variable if TRUE the flow normalized line is plotted, if FALSE not plotted}

\item{plotAnnual}{logical variable if TRUE annual flux points from WRTDS output are plotted, if FALSE not plotted}

\item{plotGenFlux}{logical variable. If \code{TRUE}, annual flux points from WRTDS_K output are plotted, if \code{FALSE} not plotted}

\item{tinyPlot}{logical variable, if TRUE plot is designed to be plotted small, as a part of a multipart figure, default is FALSE}

\item{col}{color of points on plot, see ?par 'Color Specification', default = "black"}

\item{col.pred}{color of flow normalized line on plot, see ?par 'Color Specification', default = "green"}

\item{col.gen}{color of points for WRTDS_K output on plot, see ?par 'Color Specification', default = "red"}

\item{cex}{numerical value giving the amount by which plotting symbols should be magnified}

\item{cex.axis}{magnification to be used for axis annotation relative to the current setting of cex}

\item{cex.main}{magnification to be used for main titles relative to the current setting of cex}

\item{lwd}{number line width, default = 2.}

\item{customPar}{logical defaults to FALSE. If TRUE, par() should be set by user before calling this function 
(for example, adjusting margins with par(mar=c(5,5,5,5))). If customPar FALSE, EGRET chooses the best margins depending on tinyPlot.}

\item{\dots}{arbitrary graphical parameters that will be passed to genericEGRETDotPlot function (see ?par for options)}
}
\description{
Data come from named list, which contains a Daily dataframe with the daily flow data,
and an INFO dataframe with metadata.

The annual results reported are for a specified "period of analysis" which can be 
an entire water year, a calendar, a season or even an individual month. 
The user specifies this period of analysis in the call to setupYears.
Note that when a period of analysis less than a full year is used
the reported results can be confusing.  The flux values are average flux rates
for the specified part of the year.  They are not a mass.  Rather, they are a mass per unit time.  Thus, it
is common for the flux for a specific period of analysis within a given year to be greater than the
flux for that entire year.

Three versions of annual flux can be plotted
 "Annual" version is the mean flux computed directly from the WRTDS model
 "GenConc" version uses the WRTDS_K calculation, that uses an auto-regressive formulation to improve the accuracy of the mean flux.  It has been shown to be more accurate than the "Annual" version.
 "FlowNormalized" version eliminates the interannual variability by integrating the WRTDS model results over the full probability distribution of discharge
 See introduction to the EGRET vignette for more details on these three options.

Although there are a lot of optional arguments to this function, most are set to a logical default.
}
\details{
running modelEstimation is required before running this function.
 if genFlux is plotted then running WRTDSKalman is also required.
}
\examples{
yearStart <- 2001
yearEnd <- 2010
eList <- Choptank_eList
# Water year:
\donttest{
plotFluxHist(eList)
plotFluxHist(eList, yearStart, yearEnd, fluxUnit = 1)
plotFluxHist(eList, yearStart, yearEnd, fluxUnit = 'kgDay')
# Graphs consisting of Jun-Aug
eList <- setPA(eList, paStart=6,paLong=3)
plotFluxHist(eList) #' 
}
}
\seealso{
\code{\link{setupYears}}
}
\keyword{graphics}
\keyword{statistics}
\keyword{water-quality}
