% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populateDaily.r
\name{populateDaily}
\alias{populateDaily}
\title{Populate Daily data frame}
\usage{
populateDaily(rawData, qConvert, verbose = TRUE, adjust = TRUE,
  fill = FALSE, maxgap = 21)
}
\arguments{
\item{rawData}{dataframe contains at least Date, Q, Qualifier columns.}

\item{qConvert}{numeric conversion to cubic meters per second.}

\item{verbose}{logical specifying whether or not to display messages.}

\item{adjust}{logical specifying whether or not to add a constant to zero values
to allow log transformation. Defaults to TRUE.}

\item{fill}{logical specifying whether to fill NA values by linear interpolation.
Defaults to FALSE.}

\item{maxgap}{Maximum number of NA days allowed for interpolating gaps.
Default is 21. Only used if fill is set to TRUE.}
}
\value{
A data frame 'Daily' with the following columns:
\tabular{lll}{
Name \tab Type \tab Description \cr
Q \tab numeric \tab Discharge in m^3/s\cr
Julian \tab integer \tab Number of days since Jan. 1, 1850\cr
Month \tab integer \tab Month of the year [1-12] \cr
Day \tab integer \tab Day of the year [1-366] \cr
DecYear \tab numeric \tab Decimal year \cr
MonthSeq \tab integer \tab Number of months since January 1, 1850 \cr
Qualifier \tab character \tab Qualifying code \cr
i \tab integer \tab Index of days, starting with 1 \cr
LogQ \tab numeric \tab Natural logarithm of Q  \cr
Q7 \tab numeric \tab 7 day running average of Q \cr
Q30 \tab numeric \tab 30 day running average of Q \cr
}
}
\description{
Using a data frame that has at least Date, Q, Qualifier, populates the rest of the basic Daily data frame used in EGRET analysis.
}
\examples{
Date <- as.character(seq(as.Date("2001/1/1"),
         as.Date("2002/1/2"), by = "day"))
Q <- c(-1:365)
Qualifier <- rep("",367)
dataInput_complete <- data.frame(Date, Q, Qualifier, stringsAsFactors=FALSE)
dataInput <- dataInput_complete[-4:-5,]
Daily <- populateDaily(dataInput, 1)
Daily_fill <- populateDaily(dataInput, 1, fill = TRUE)

dataInput <- dataInput_complete[-4:-35,]
dataInput <- dataInput[-200:-255,]
Daily_fill2 <- populateDaily(dataInput, 1, fill = TRUE)
plot(Daily_fill2$Date, Daily_fill2$Q)

eList <- Choptank_eList
Daily_chop <- eList$Daily
df <- Daily_chop[,c("Date", "Q")]
df <- df[-2:-5, ]
df <- df[-100:-200,]
D2 <- populateDaily(df, 1, fill = TRUE)
plot(D2$Date[1:10], D2$Q[1:10])
plot(D2$Date[1:365], D2$Q[1:365])
plot(D2$Date, D2$Q)

df <- df[-5000:-10000, ]
D3 <- populateDaily(df, 1, fill = TRUE)
plot(D3$Date, D3$Q)
}
\seealso{
\code{\link{readNWISDaily}}, \code{\link{readUserDaily}}
}
\author{
Robert M. Hirsch \email{rhirsch@usgs.gov}
}
\keyword{WRTDS}
\keyword{flow}
