% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_functions.R
\name{populateDailySamp}
\alias{populateDailySamp}
\title{populateDailySamp}
\usage{
populateDailySamp(eList)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes and surfaces matrix.}
}
\description{
Creates a new Daily data frame that has appended to it the trueFlux for the day and the standardized residual
For the day on all sampled days.
}
\keyword{internal}
