% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populateParameterINFO.r
\name{populateParameterINFO}
\alias{populateParameterINFO}
\title{Populate Parameter Information Columns}
\usage{
populateParameterINFO(parameterCd, INFO, interactive = TRUE)
}
\arguments{
\item{parameterCd}{character USGS parameter code}

\item{INFO}{dataframe with value and code columns. Default is INFO}

\item{interactive}{logical Option for interactive mode.  If TRUE, there is user interaction for error handling
and data checks. Default is TRUE. If running in batch, should be set to FALSE.}
}
\value{
INFO dataframe
}
\description{
Populates INFO data frame with additional user-supplied information concerning the measured parameter.
}
\examples{
\donttest{

INFO <- dataRetrieval::read_waterdata_monitoring_location('USGS-01594440')
parameterCd <- "01075"
parameterData <- dataRetrieval::read_waterdata_parameter_codes(parameterCd)
INFO$param.nm <- parameterData$parameter_name
INFO$param.units <- parameterData$unit_of_measure
INFO$paramShortName <- parameterData$parameter_name
INFO$paramNumber <- parameterData$parameter_code

INFO <- populateParameterINFO(parameterCd, INFO, interactive = FALSE)
}
}
