% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeReport.r
\name{print.egret}
\alias{print.egret}
\title{Prints EGRET object}
\usage{
\method{print}{egret}(x, ...)
}
\arguments{
\item{x}{EGRET object}

\item{\dots}{additional parameters}
}
\value{
logical
}
\description{
Print function for EGRET object
}
\examples{
eList <- Choptank_eList
eList
}
\keyword{USGS}
\keyword{WRTDS}
\keyword{data}
\keyword{import}

