\name{printFluxUnitCheatSheet}
\alias{printFluxUnitCheatSheet}
\title{Available Flux Unit properties}
\usage{
  printFluxUnitCheatSheet()
}
\description{
  Cheat sheet to print out pre-defined flux unit properties
  from fluxUnit class
}
\examples{
printFluxUnitCheatSheet()
}
\keyword{graphics}
\keyword{statistics}
\keyword{water-quality}

