% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGroups.R
\name{printGroups}
\alias{printGroups}
\title{Print information about group analysis}
\usage{
printGroups(eList, groupResults)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{groupResults}{output of \code{runGroups}.}
}
\value{
text to console
}
\description{
Prints the information from the \code{runGroups} function.
This could be used to save the output to a text file.
}
\examples{
eList <- Choptank_eList

\donttest{
groupOut_1 <- runGroups(eList,  windowSide = 0,
                        group1firstYear = 1980, group1lastYear = 1990,
                        group2firstYear = 1995, group2lastYear = 2005)
                       
printGroups(eList, groupOut_1)
}

}
