\name{printqUnitCheatSheet}
\alias{printqUnitCheatSheet}
\title{Available Flow Unit properties}
\usage{
  printqUnitCheatSheet()
}
\description{
  Cheat sheet to print out pre-defined qUnit properties
  from qUnit class.
}
\examples{
printqUnitCheatSheet()
}
\keyword{graphics}
\keyword{statistics}
\keyword{water-quality}

