% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{qConst}
\alias{qConst}
\title{Flow units included in EGRET}
\description{
Flow units included:
\tabular{llll}{
Number \tab ObjectName \tab shortName \tab unitFactor \cr
1      \tab cfs      \tab Cubic Feet per Second   \tab 35.31467 \cr
2      \tab cms      \tab Cubic Meters per Second  \tab 1  \cr
3      \tab thousandCfs \tab Thousand Cubic Feet per Second   \tab  0.03531467 \cr
4      \tab thousandCms \tab Thousand Cubic Meters per Second \tab 0.001 \cr
5      \tab mmDay \tab mm per day \tab   \cr
6      \tab mmYear \tab mm per year \tab   \cr
}
}

