% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fluxUnit.R
\docType{class}
\name{qUnit-class}
\alias{qUnit-class}
\title{qUnit class}
\description{
Some details about the qUnit class
}
\details{
\describe{
   \item{qshortName}{A character specifying the short name.}

\item{qUnitFactor}{A numeric representing the conversion factor}

\item{qUnitName}{A character specifying the full name.}

\item{qUnitExpress}{An expression specifying the full name.}

\item{qUnitTiny}{An expression specifying the abbreviated name.}

\item{shortCode}{A number for quick lookup}
 }
}

