% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_functions.R
\name{randomSubset}
\alias{randomSubset}
\title{randomSubset}
\usage{
randomSubset(df, colName)
}
\arguments{
\item{df}{data frame. Must include a column named by the argument colName.}

\item{colName}{column name to check for duplicates}
}
\description{
Calculates a random subset of the data based on repeated values from
a specified column.
}
\examples{
df <- data.frame(Julian = c(1,2,2,3,4,4,4,6),
                 y = 1:8)
df
df_random <- randomSubset(df, "Julian")
df_random
}
