% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readMetaData.r
\name{readNWISInfo}
\alias{readNWISInfo}
\title{Import Metadata for USGS Data}
\usage{
readNWISInfo(siteNumber, parameterCd, interactive = TRUE)
}
\arguments{
\item{siteNumber}{character USGS site number.  This is usually an 8 digit number}

\item{parameterCd}{character USGS parameter code.  This is usually an 5 digit number.}

\item{interactive}{logical Option for interactive mode.  If true, there is user interaction for error handling and data checks.}
}
\value{
INFO dataframe with at least param.nm, param.units, parameShortName, paramNumber
}
\description{
Populates INFO data frame for EGRET study.  If either station number or parameter code supplied, imports data about a particular USGS site from NWIS web service.
This function gets the data from here: \url{http://waterservices.usgs.gov/}
If either station number or parameter code is not supplied, the user will be asked to input data.
Additionally, the user will be asked for:
staAbbrev - station abbreviation, will be used in naming output files and for structuring batch jobs
constitAbbrev - constitute abbreviation
}
\examples{
# These examples require an internet connection to run
# Automatically gets information about site 05114000 and temperature, no interaction with user
INFO <- readNWISInfo('05114000','00010',interactive=FALSE)
}
\keyword{USGS}
\keyword{data}
\keyword{import}
\keyword{service}
\keyword{web}

