\docType{data}
\name{RioDaily}
\alias{RioDaily}
\title{Example Daily dataframe for Rio Grande}
\description{
  Example data representing streamflowfrom the Rio Grande
  at Embudo, NM, USGS data.
}
\author{
  Robert M. Hirsch \email{rhirsch@usgs.gov}
}
\keyword{data}
\keyword{datasets}
\keyword{quality}
\keyword{streamflow}
\keyword{water}

