\docType{data}
\name{RioINFO}
\alias{RioINFO}
\title{Example INFO dataframe for Rio Grande River}
\description{
  Example data representing meta-data from the Rio Grande
  at Embudo, NM, USGS data
}
\author{
  Robert M. Hirsch \email{rhirsch@usgs.gov}
}
\keyword{data}
\keyword{datasets}
\keyword{quality}
\keyword{water}

