% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveResults.R
\name{saveResults}
\alias{saveResults}
\title{A utility program for saving the contents of the workspace
  
 This function saves the workspace. Future versions of EGRET will not include this function,
 use saveRDS to save individual eList objects.
 It assigns the file a name using the abbreviations for station and constituent.}
\usage{
saveResults(savePath, eList)
}
\arguments{
\item{savePath}{character specifying the full pathname of the folder where the file is to be saved ending with the final slash}

\item{eList}{named list with at least the INFO dataframe}
}
\description{
A utility program for saving the contents of the workspace
  
 This function saves the workspace. Future versions of EGRET will not include this function,
 use saveRDS to save individual eList objects.
 It assigns the file a name using the abbreviations for station and constituent.
}
\keyword{statistics}
\keyword{water-quality}
