% Generated by roxygen2 (4.0.2): do not edit by hand
\name{selectDays}
\alias{selectDays}
\title{Creates a subset of the Daily data frame that only contains data for the specified period of analysis}
\usage{
selectDays(df, paLong, paStart)
}
\arguments{
\item{df}{dataframe with Q}

\item{paLong}{a numeric value for the length of the period of Analysis, must be an integer from 1 to 12}

\item{paStart}{a numeric value for the starting month of the period of analysis, must be an integer from 1 to 12}
}
\value{
localDaily a data frame containing the daily data but only for the period of analysis (not all months)
}
\description{
Function is not called by user, but is called by plotSDLogQ
}
\examples{
Daily <- ChopDaily
DailySubset <- selectDays(eList, 4, 11)
}
\keyword{statistics}
\keyword{streamflow}

