% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startEnd.R
\name{startEnd}
\alias{startEnd}
\title{startEnd}
\usage{
startEnd(paStart, paLong, year)
}
\arguments{
\item{paStart}{numeric integer specifying the starting month for the period of analysis, 1<=paStart<=12, default is 10}

\item{paLong}{numeric integer specifying the length of the period of analysis, in months, 1<=paLong<=12, default is 12}

\item{year}{integer year}
}
\value{
Date list
}
\description{
Returns two date variables for a combination of paStart, paLong, and year
}
\examples{
paStart <- 10
paLong <- 12
year <- 1999
startEnd(paStart, paLong, year)
}
