% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch}
\alias{stitch}
\title{stitch}
\usage{
stitch(eList, sample1StartDate, sample1EndDate, sample2StartDate,
  sample2EndDate, surfaceStart = NA, surfaceEnd = NA, minNumObs = 100,
  minNumUncen = 50, windowY = 7, windowQ = 2, windowS = 0.5,
  edgeAdjust = TRUE, verbose = FALSE, run.parallel = FALSE)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{sample1StartDate}{The Date (or character in YYYY-MM-DD) of the first sample to be used (if NA it is the first Date in eList$Sample)}

\item{sample1EndDate}{The Date (or character in YYYY-MM-DD) of the last sample to be used in the first section}

\item{sample2StartDate}{The Date (or character in YYYY-MM-DD) of the first sample to be used in the second section}

\item{sample2EndDate}{The Date (or character in YYYY-MM-DD) of the last sample to be used in the second section}

\item{surfaceStart}{The Date (or character in YYYY-MM-DD) that is the start of the WRTDS model to be estimated and the last of the 
daily outputs to be generated (if NA it is sampleStartDate)}

\item{surfaceEnd}{The Date (or character in YYYY-MM-DD) that is the end of the WRTDS model to be estimated and the last of the daily outputs 
to be generated (if NA it is sampleEndDate)}

\item{minNumObs}{numeric specifying the miniumum number of observations required to run the weighted regression, default is 100}

\item{minNumUncen}{numeric specifying the minimum number of uncensored observations to run the weighted regression, default is 50}

\item{windowY}{numeric specifying the half-window width in the time dimension, in units of years, default is 7}

\item{windowQ}{numeric specifying the half-window width in the discharge dimension, units are natural log units, default is 2}

\item{windowS}{numeric specifying the half-window with in the seasonal dimension, in units of years, default is 0.5}

\item{edgeAdjust}{logical specifying whether to use the modified method for calculating the windows at the edge of the record.
The modified method tends to reduce curvature near the start and end of record.  Default is TRUE.}

\item{paLong}{numeric integer specifying the length of the period of analysis, in months, 1<=paLong<=12, default is 12}

\item{paStart}{numeric integer specifying the starting month for the period of analysis, 1<=paStart<=12, default is 10}
}
\description{
This function creates a continuous surfaces object that starts just before 
surfaceStart and ends just after surfaceEnd.
}
\details{
It has a wall in it and the wall comes just after lastDaySample1
Arguments 2 through 5 are all dates, but just to make sure we will do a conversion to as.Date
the only use being made of eList$Daily is to establish the minimum and maximum discharges for the surfaces
the eList$Sample will be split out into Sample1 and Sample2 (there will be no overlap)
}
\examples{
eList <- Choptank_Phos

surfaceStart <- "1986-10-01"
surfaceEnd <- "2012-09-30"

# Surface skips a few years:
sample1StartDate <- "1986-10-01"
sample1EndDate <- "1992-09-30"
sample2StartDate <- "1996-10-01"
sample2EndDate <- "2012-09-30"

\dontrun{
surface_skip <- stitch(eList, 
                         sample1StartDate, sample1EndDate, 
                         sample2StartDate, sample2EndDate,
                         surfaceStart, surfaceEnd)

# Surface overlaps a few years:
sample1StartDate <- "1986-10-01"
sample1EndDate <- "1996-09-30"
sample2StartDate <- "1992-10-01"
sample2EndDate <- "2012-09-30"

surface_overlap <- stitch(eList, 
                         sample1StartDate, sample1EndDate, 
                         sample2StartDate, sample2EndDate)
}
}
