% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexFNnew.R
\name{subFN}
\alias{subFN}
\title{subFN}
\usage{
subFN(eList, flowNormYears = "all", waterYear = TRUE)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{flowNormYears}{vector of flow years}

\item{waterYear}{logical. Should years be water years (\code{TRUE}) or calendar years (\code{FALSE})}
}
\value{
data frame in the Daily format
}
\description{
Calculates flow normalized flux and concentration with a subset of the flow record.
}
\examples{
eList <- Choptank_eList

flowNormYears <- c(1985:2002,2006:2010)
temp_daily <- subFN(eList, flowNormYears)
}
