% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startEnd.R
\name{surfaceStartEnd}
\alias{surfaceStartEnd}
\title{surfaceStartEnd}
\usage{
surfaceStartEnd(paStart, paLong, Date1, Date2)
}
\arguments{
\item{paStart}{numeric integer specifying the starting month for the period of analysis, 1<=paStart<=12, default is 10}

\item{paLong}{numeric integer specifying the length of the period of analysis, in months, 1<=paLong<=12, default is 12}

\item{Date1}{earliest date of Sample (to set surface start)}

\item{Date2}{latest date of Sample (to set surface end)}
}
\description{
surfaceStartEnd
}
\examples{
eList <- Choptank_eList
Date1 <- eList$Sample$Date[1]
Date2 <- range(eList$Sample$Date)[2]
surfaceStartEnd(10, 12, Date1, Date2)
}
