% Generated by roxygen2 (4.0.2): do not edit by hand
\name{boxConcMonth}
\alias{boxConcMonth}
\title{Box plot of the water quality data by month}
\usage{
boxConcMonth(localSample = Sample, localINFO = INFO, printTitle = TRUE,
  cex = 0.8, cex.axis = 1.1, cex.main = 1.1, las = 1,
  logScale = FALSE, tcl = 0.5, tinyPlot = FALSE, customPar = FALSE, ...)
}
\arguments{
\item{localSample}{data frame that contains the concentration data, default name is Sample}

\item{localINFO}{data frame that contains the metadata, default name is INFO}

\item{printTitle}{logical variable if TRUE title is printed, if FALSE not printed (this is best for a multi-plot figure)}

\item{cex}{numerical value giving the amount by which plotting symbols should be magnified}

\item{cex.axis}{magnification to be used for axis annotation relative to the current setting of cex}

\item{cex.main}{magnification to be used for main titles relative to the current setting of cex}

\item{las}{numeric in {0,1,2,3}; the style of axis labels, see ?par}

\item{logScale}{logical if TRUE y plotted in log axis}

\item{tcl}{number defaults to 0.5, specifies length of tick marks as fraction of height of a line of text}

\item{tinyPlot}{logical variable, if TRUE plot is designed to be plotted small as part of a multi-plot figure, default is FALSE.}

\item{customPar}{logical defaults to FALSE. If TRUE, par() should be set by user before calling this function}

\item{\dots}{arbitrary graphical parameters that will be passed to genericEGRETDotPlot function (see ?par for options)}
}
\description{
Data come from a data frame named Sample which contains the Sample Data.
The metadata come from a data frame named INFO.
These data frames must have been created by the dataRetrieval package.

 Although there are a lot of optional arguments to this function, most are set to a logical default. If your workspace
 contains an INFO and Sample dataframes, then the following R code will produce a plot:
 \code{boxConcMonth()}
}
\examples{
Sample <- ChopSample
INFO <- ChopINFO
# Water year:
boxConcMonth()
# Graphs consisting of Jun-Aug
INFO <- setPA(paStart=6,paLong=3)
boxConcMonth()
}
\keyword{graphics}
\keyword{statistics}
\keyword{water-quality}

