% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ChopDaily}
\alias{ChopDaily}
\title{Example Daily dataframe}
\format{Data frame with 11688 rows and 18 columns:
\tabular{llll}{
ColumnName \tab Type \tab Description \tab Units  \cr
Date \tab Date \tab Date \tab date \cr
Q \tab number \tab Discharge in cms \tab cms \cr
Julian \tab number \tab Number of days since January 1, 1850 \tab days \cr
Month \tab integer \tab Month of the year [1-12] \tab months \cr
Day \tab integer \tab Day of the year [1-366] \tab days \cr
DecYear \tab number \tab Decimal year \tab years \cr
MonthSeq \tab integer \tab Number of months since January 1, 1850 \tab months \cr
Qualifier \tab string \tab Qualifing code \tab character \cr
i \tab integer \tab Index of days, starting with 1 \tab days \cr
LogQ \tab number \tab Natural logarithm of Q \tab numeric \cr
Q7 \tab number \tab 7 day running average of Q \tab cms \cr
Q30 \tab number \tab 30 running average of Q \tab cms \cr
yHat \tab number \tab The WRTDS estimate of the log of concentration \tab numeric \cr
SE \tab number \tab The WRTDS estimate of the standard error of yHat \tab numeric \cr
ConcDay \tab number \tab The WRTDS estimate of concentration \tab mg/L \cr
FluxDay \tab number \tab The WRTDS estimate of flux \tab kg/day \cr
FNConc \tab number \tab Flow normalized estimate of concentration \tab mg/L \cr
FNFlux \tab number \tab Flow Normalized estimate of flux \tab kg/day \cr
}}
\source{
Data retrieved from NWIS water services (\url{http://waterservices.usgs.gov/}) via the EGRET R package.
The USGS site id is 01491000. The period requested was 1979-10-01 to 2011-09-30.
}
\description{
Example data representing streamflowfrom the Choptank River at Greensboro, MD,  USGS data
as agumented by the modelEstimation process.
}
\examples{
Daily <- ChopDaily
with(Daily, plot(Date, LogQ, type="l"))
}
\author{
Robert M. Hirsch \email{rhirsch@usgs.gov}
}
\keyword{data}
\keyword{quality}
\keyword{streamflow}
\keyword{water}

