% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EGRET.R
\docType{data}
\name{Constants}
\alias{Constants}
\alias{fluxConst}
\alias{monthInfo}
\alias{qConst}
\title{Constants included with EGRET}
\description{
\itemize{
 \item{fluxConst}{Flux conversion object}
 \item{qConst}{Flow conversion object}
 \item{monthInfo}{Month object}
}
}
\examples{
fluxConst
fluxConst[['kgDay']]
fluxConst[['kgDay']]@unitName
qConst
qConst[['cfs']]
qConst[['cfs']]@qUnitName
}
\keyword{datasets}

