% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{decimalHighLow}
\alias{decimalHighLow}
\title{decimalHighLow}
\usage{
decimalHighLow(df)
}
\arguments{
\item{df}{data.frame with Date, DecYear, and Month columns}
}
\value{
list with DecHigh and DecLow (water year high/low decimal values)
}
\description{
decimalHighLow
}
\examples{
eList <- Choptank_Phos
highLow <- decimalHighLow(eList$Sample)

DecHigh <- highLow[["DecHigh"]]
DecLow <- highLow[["DecLow"]]
}
