\docType{package}
\name{EGRET-package}
\alias{EGRET-package}
\title{EGRET package includes WRTDS and flowHistory}
\description{
\tabular{ll}{ Package: \tab EGRET\cr Type: \tab Package\cr
Version: \tab 1.2.4\cr Date: \tab 2013-06-28\cr License:
\tab Unlimited for this package, dependencies have more
restrictive licensing.\cr Copyright: \tab This software is
in the public domain because it contains materials that
originally came from the United States Geological Survey,
an agency of the United States Department of Interior. For
more information, see the official USGS copyright policy at
http://www.usgs.gov/visual-id/credit_usgs.html#copyright\cr
LazyLoad: \tab yes\cr }
}
\details{
Collection of functions to do WRTDS and flowHistory
analysis, and produce graphs and tables of data and results
from these analyses.
}
\author{
Robert M. Hirsch \email{rhirsch@usgs.gov}, Laura De Cicco
\email{ldecicco@usgs.gov}
}
\references{
Hirsch, R. M., Moyer, D. L. and Archfield, S. A. (2010),
Weighted Regressions on Time, Discharge, and Season
(WRTDS), with an Application to Chesapeake Bay River
Inputs. JAWRA Journal of the American Water Resources
Association, 46: 857-880. doi:
10.1111/j.1752-1688.2010.00482.x
}
\keyword{graphics}
\keyword{statistics}
\keyword{streamflow}
\keyword{water-quality}

