% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EGRET.R
\docType{package}
\name{EGRET-package}
\alias{EGRET-package}
\title{EGRET package includes WRTDS and flowHistory}
\description{
\tabular{ll}{
Package: \tab EGRET\cr
Type: \tab Package\cr
License: \tab Unlimited for this package, dependencies have more restrictive licensing.\cr
Copyright: \tab This software is in the public domain because it contains materials
that originally came from the United States Geological Survey, an agency of
the United States Department of Interior. For more information, see the
official USGS copyright policy at
http://www.usgs.gov/visual-id/credit_usgs.html#copyright\cr
LazyLoad: \tab yes\cr
}
}
\details{
Collection of functions to do WRTDS and flowHistory analysis,
 and produce graphs and tables of data and results from these analyses.
}
\author{
Robert M. Hirsch \email{rhirsch@usgs.gov}, Laura De Cicco \email{ldecicco@usgs.gov}
}
\references{
Hirsch, R.M., and De Cicco, L.A., 2014, User guide to Exploration and Graphics for RivEr Trends
(EGRET) and dataRetrieval: R packages for hydrologic data: U.S. Geological Survey Techniques and Methods book 4,
chap. A10, 94 p., \url{http://dx.doi.org/10.3133/tm4A10}
}
\keyword{graphics}
\keyword{statistics}
\keyword{streamflow}
\keyword{water-quality}

