% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexFNnew.R
\name{estFNsegs}
\alias{estFNsegs}
\title{Segment estimates for flow normalization}
\usage{
estFNsegs(eList, dateInfo)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{dateInfo}{dataframe with sampleSegStart, flowSegStart, flowSegEnd, sampleSegEnd}
}
\description{
Segment estimates for flow normalization
}
