\docType{data}
\name{exINFOEnd}
\alias{exINFOEnd}
\title{Example final INFO dataframe}
\description{
  Example data representing meta-data from the Choptank
  River at Greensboro, MD, USGS data as augmented by the
  functions setupYears (for WRTDS) and setPA (for
  flowHistory)
}
\author{
  Robert M. Hirsch \email{rhirsch@usgs.gov}
}
\keyword{data}
\keyword{datasets}
\keyword{quality}
\keyword{water}

