\docType{data}
\name{exINFOStart}
\alias{exINFOStart}
\title{Example initial INFO dataframe}
\description{
  Example data representing meta-data from the Choptank
  River at Greensboro, MD, USGS data This is the data that
  is initially generated from the dataRetrieval package.
}
\author{
  Robert M. Hirsch \email{rhirsch@usgs.gov}
}
\keyword{data}
\keyword{datasets}
\keyword{quality}
\keyword{water}

