% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{exsurfaces}
\alias{exsurfaces}
\title{Example surface array}
\description{
Example data representing surfaces, from the fitted model
of Nitrate data for the Choptank River at Greensboro, MD,  USGS data
output from the function estSurfaces
}
\author{
Robert M. Hirsch \email{rhirsch@usgs.gov}
}
\keyword{data}
\keyword{quality}
\keyword{water}

