% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexFNnew.R
\name{flexFN}
\alias{flexFN}
\title{Flexible Flow Normalization}
\usage{
flexFN(eList, dateInfo, sampleStart = "sampleSegStart",
  flowStart = "flowSegStart", flowEnd = "flowSegEnd")
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{dateInfo}{data frame with 3 columns, their names defined by sampleStart, flowStart, flowEnd}

\item{sampleStart}{integer vector of start years (water) for each FN conc/flux segment}

\item{flowStart}{integer vector of start years (water) for flow normalization}

\item{flowEnd}{integer vector of end years (water) for flow normalization}
}
\description{
Flexible Flow Normalization
}
\examples{
eList <- Choptank_eList
eList <- setUpEstimation(eList)
sampleSegStart <- c(1980,1990,2000)
flowSegStart <- c(1980,1985,1992)
flowSegEnd <- c(1994,2004,2011)
dateInfo <- data.frame(sampleSegStart, 
                       flowSegStart, 
                       flowSegEnd)
\dontrun{
eList <- flexFN(eList, dateInfo)
plotFluxHist(eList)

eList <- Choptank_eList
eList <- setUpEstimation(eList)
sampleSegStart <- c(1980,1985,2000)
flowSegStart <- c(1980,1990,2000)
flowSegEnd <- c(1990,2000,2010)
dateInfo <- data.frame(sampleSegStart, 
                       flowSegStart, 
                       flowSegEnd)
eList <- flexFN(eList, dateInfo)
plotFluxHist(eList)
}
}
