% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexFN.R
\name{flexPlotAddOn}
\alias{flexPlotAddOn}
\title{Flexible Flow Normalization Plot Add On}
\usage{
flexPlotAddOn(eList, showArrows = TRUE, showRect = TRUE,
  customPalette = NULL)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{showArrows}{logical whether or not to show arrows representing flow segments}

\item{showRect}{logical whether or not to show rectangles representing sample segments}

\item{customPalette}{character vector of colors as a hexadecimal string of the form "#rrggbb". 
Defaults to NULL, which indicates the use of a default palette (up to 21 segments).}
}
\description{
Flexible Flow Normalization Plot Add On
}
\examples{
eList <- Choptank_eList
eList <- setUpEstimation(eList)
sampleStart <- c("1980-01-01","1990-01-01","2000-01-01")
sampleEnd <- c("1995-06-06","2002-09-30","2011-02-01")
flowStart <- c("1980-02-01","1985-06-01","1992-10-10")
flowEnd <- c("1994-06-15","2004-03-03","2011-02-01")
dateInfo <- data.frame(sampleStart,
                       sampleEnd,
                       flowStart, 
                       flowEnd, stringsAsFactors = FALSE)
\dontrun{
newEList <- flexFN(eList, dateInfo)
plotFluxHist(newEList)
flexPlotAddOn(newEList)

plotFluxHist(newEList)
flexPlotAddOn(eList, customPalette=c("#d5ce48", "#fd300f", "#3e0289"))
}
}
