% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fluxUnit.R
\docType{class}
\name{fluxUnit-class}
\alias{fluxUnit-class}
\title{fluxUnit class}
\description{
Some details about the fluxUnit class
}
\details{
\describe{
   \item{shortName}{A character specifying the short name.}

\item{unitFactor}{A numeric representing the conversion factor}

\item{unitName}{A character specifying the full name.}

\item{unitExpress}{An expression specifying the full name starting with Observed.}

\item{unitExpressTiny}{An expression specifying the abbreviated name starting with Observed.}

\item{unitEstimate}{An expression specifying the full name starting with Estimated.}

\item{unitEstimateTiny}{An expression specifying the abbreviated name starting with Estimated.}

\item{unitUSGS}{A character specifying flux with full text.}

\item{shortCode}{A number for quick lookup}
 }
}

