% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generalAxis}
\alias{generalAxis}
\title{Axis generation for log discharge}
\usage{
generalAxis(x, maxVal, minVal, logScale = FALSE, tinyPlot = FALSE,
  padPercent = 5, concentration = TRUE, localINFO = INFO)
}
\arguments{
\item{x}{vector to create scale about}

\item{maxVal}{number maximum value on returned scale}

\item{minVal}{number minimum value on returned scale}

\item{logScale}{logical whether or not to return a log scale}

\item{tinyPlot}{logical}

\item{padPercent}{number used to pad the max and min if not specified}

\item{concentration}{logical if concentration=TRUE, labels returned as concentration units, otherwise flux units.}

\item{localINFO}{data frame that contains the metadata, default name is INFO}
}
\description{
Discharge axis tick generation
}
\examples{
Daily <- ChopDaily
INFO <- ChopINFO
x <- Daily$Q
max <- max(x)
min <- 0
generalAxis(x, max, min)
min <- min(x)
generalAxis(x, max, min, log=TRUE)
}
\keyword{graphics}
\keyword{statistics}
\keyword{water-quality}

