% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_functions.R
\name{genmissing}
\alias{genmissing}
\title{genmissing}
\usage{
genmissing(X1, XN, rho, N)
}
\arguments{
\item{X1}{value before the gap}

\item{XN}{value after the gap}

\item{rho}{the lag one autocorrelation}

\item{N}{the length of the sequence including X1 and XN. It
is two more than the gap length}
}
\description{
Generates a single realization of an autoregression lag-one time series through a data gap
Conditioned on a fixed value just prior to the gap (X1) and a fixed value just after the gap (XN)
}
\details{
Written by Tim Cohn
}
\keyword{internal}
