% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egret_objects.R
\name{getDaily}
\alias{getDaily}
\alias{getDaily.egret}
\alias{getDaily.default}
\title{Get Daily dataframe from EGRET object}
\usage{
getDaily(x, ...)

\method{getDaily}{egret}(x, ...)

\method{getDaily}{default}(x, ...)
}
\arguments{
\item{x}{EGRET object or named list}

\item{\dots}{additional parameters}
}
\value{
Daily dataframe
}
\description{
From a named list or EGRET object, extract the Daily dataframe
}
\examples{
eList <- Choptank_eList
Daily <- getDaily(eList)
}
\seealso{
\code{\link{readNWISDaily}}, \code{\link{readNWISSample}}
}
\keyword{USGS}
\keyword{WRTDS}
\keyword{data}
\keyword{import}
