% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeAnnualSeries}
\alias{makeAnnualSeries}
\title{Produces annual series of 8 streamflow statistics (and a lowess smooth of them) from daily streamflow data}
\usage{
makeAnnualSeries(localDaily = Daily, localINFO = INFO, edgeAdjust = TRUE)
}
\arguments{
\item{localDaily}{data frame that contains the daily streamflow data}

\item{localINFO}{data frame that contains the metadata}

\item{edgeAdjust}{logical specifying whether to use the modified method for calculating the windows at the edge of the record.  The modified method tends to reduce curvature near the start and end of record.  Default is TRUE.}
}
\value{
annualSeries data frame that contains the annual series of streamflow statistics
}
\description{
Part of the flowHistory system.  The data come from Daily and INFO data frames.
Note that the function setPA must be run before this to establish the period of analysis (e.g. water year).
}
\details{
\tabular{ll}{
istat  \tab Name  \cr
1 \tab 1-day minimum discharge \cr
2 \tab 7-day minimum discharge \cr
3 \tab 30-day minimum discharge \cr
4 \tab median discharge \cr
5 \tab mean discharge \cr
6 \tab 30-day maximum discharge \cr
7 \tab 7-day maximum discharge \cr
8 \tab 1-day maximum discharge \cr
}
}
\examples{
Daily <- ChopDaily
INFO <- ChopINFO
annualSeries <- makeAnnualSeries()
}
\keyword{statistics}
\keyword{streamflow}
\keyword{trends}

