% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot15.R
\name{plot15}
\alias{plot15}
\title{Makes 15 graphs of streamflow statistics on a single page}
\usage{
plot15(eList, yearStart, yearEnd)
}
\arguments{
\item{eList}{named list with at least the Daily and INFO dataframes}

\item{yearStart}{A numeric value for year in which the graph should start, default is NA, which indicates that the graph should start with first annual value}

\item{yearEnd}{A numeric value for year in which the graph should end, default is NA, which indicates that the graph should end with last annual value}
}
\description{
Part of flowHistory system.
}
\examples{
eList <- Choptank_eList
\dontrun{
pdf("plot15.pdf",heigh=10,width=8)
plot15(eList, yearStart=1990,yearEnd=2000)
dev.off()
}
}
\seealso{
\code{\link{plot1of15}}
}
\keyword{graphics}
\keyword{statistics}
\keyword{streamflow}

