% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMonthTrend.R
\name{plotMonthTrend}
\alias{plotMonthTrend}
\title{Plot monthly trend result from runPairs}
\usage{
plotMonthTrend(pairResults, yMax = NA, arrowFactor = 1, flux = TRUE,
  printTitle = TRUE)
}
\arguments{
\item{pairResults}{results from \code{runPairs}.}

\item{yMax}{numeric. Upper limit for plot. Default is \code{NA},
which will use the maximum of the data.}

\item{arrowFactor}{numeric. Scaling factor for the size of the arrows.
The arrows are automatically scaled to the overall trend. This scaling 
factor helps adjust how big/small they are.}

\item{flux}{logical. \code{TRUE} is flux, \code{FALSE} is concentration.
Default is \code{TRUE}.}
}
\description{
Plot monthly trend result from runPairs
}
\examples{

eList <- Choptank_eList
year1 <- 1985
year2 <- 2010

\donttest{

pairOut_1 <- runPairs(eList, year1, year2, windowSide = 0)

plotMonthTrend(pairOut_1)
plotMonthTrend(pairOut_1, flux = FALSE)

eList <- setPA(eList, paStart = 12, paLong = 3)
pairOut_2 <- runPairs(eList, year1, year2, windowSide = 0)

plotMonthTrend(pairOut_2)

eList <- setPA(eList, paStart = 1, paLong = 12)
pairOut_3 <- runPairs(eList, year1, year2, windowSide = 0)

plotMonthTrend(pairOut_3)

}
 
}
