% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populateDaily.r
\name{populateDaily}
\alias{populateDaily}
\title{Populate Daily data frame}
\usage{
populateDaily(rawData, qConvert, verbose = TRUE, adjust = TRUE,
  fill = FALSE, maxgap = 21, fill_type = "interpolation")
}
\arguments{
\item{rawData}{dataframe contains at least Date, Q, Qualifier columns.}

\item{qConvert}{numeric conversion to cubic meters per second.}

\item{verbose}{logical specifying whether or not to display messages.}

\item{adjust}{logical specifying whether or not to add a constant to zero values
to allow log transformation. Defaults to TRUE.}

\item{fill}{logical specifying whether to fill NA values by linear interpolation.
Defaults to FALSE.}

\item{maxgap}{Maximum number of NA days allowed for interpolating gaps.
Default is 21. Only used if fill is set to TRUE.}

\item{fill_type}{character to define what process to fill missing data. Options are
"interpolation" - linear interpolation from the
`zoo::na.approx`, or "log_interp" - linear interpolation in the log space.
Only used if fill is set to TRUE.}
}
\value{
A data frame 'Daily' with the following columns:
\tabular{lll}{
Name \tab Type \tab Description \cr
Q \tab numeric \tab Discharge in m^3/s\cr
Julian \tab integer \tab Number of days since Jan. 1, 1850\cr
Month \tab integer \tab Month of the year [1-12] \cr
Day \tab integer \tab Day of the year [1-366] \cr
DecYear \tab numeric \tab Decimal year \cr
MonthSeq \tab integer \tab Number of months since January 1, 1850 \cr
Qualifier \tab character \tab Qualifying code \cr
i \tab integer \tab Index of days, starting with 1 \cr
LogQ \tab numeric \tab Natural logarithm of Q  \cr
Q7 \tab numeric \tab 7 day running average of Q \cr
Q30 \tab numeric \tab 30 day running average of Q \cr
}
}
\description{
Using a data frame that has at least Date, Q, Qualifier, populates the rest of the basic Daily data frame used in EGRET analysis.
}
\examples{
Date <- as.character(seq(from = as.Date("2001/1/1"),
                         to = as.Date("2002/1/2"),
                         by = "day"))
Q <- c(-1:365)
Qualifier <- rep("",367)
dataInput_complete <- data.frame(Date, Q, Qualifier)
dataInput <- dataInput_complete[-4:-5,]

# No fill, but with 0 and negative:
Daily <- populateDaily(dataInput, qConvert = 1)

# No negatives/zeros:
Q <- 2+sin(seq(from = 0, to = 2*pi, length.out = 367))
Q <- jitter(Q, factor = 500)
plot(Q, ylim = c(0, 3.2))
dataInput_complete <- data.frame(Date, Q, Qualifier)
# Remove some rows to test missing:
dataInput <- dataInput_complete[-4:-5,]
dataInput <- dataInput[-10:-20,]

# No fill:
Daily <- populateDaily(dataInput, qConvert = 1)
plot(Daily$Date[1:30], Daily$Q[1:30], type = "b", ylim = c(0, 3.2))

# Linear interpolation:
Daily_fill <- populateDaily(dataInput,
                            qConvert = 1,
                            fill = TRUE,
                            fill_type = "interpolation")
plot(Daily_fill$Date[1:30],
     Daily_fill$Q[1:30],
     col = as.factor(Daily_fill$Qualifier[1:30]),
     type = "b", pch = 16, ylim = c(0, 3.2),
     main = "Linear Interpolation")


# Add a gap that is too big do deal with:
dataInput <- dataInput_complete[-4:-20,]
dataInput <- dataInput[-200:-255,]

Daily_interp <- populateDaily(dataInput,
                              qConvert = 1,
                              fill = TRUE,
                              fill_type = "interpolation")
plot(Daily_interp$Date, Daily_interp$Q,
     col = as.factor(Daily_interp$Qualifier),
     main = "Linear Interpolation",
     type = "b", pch = 16, ylim = c(0, 3.2))
plot(Daily_interp$Date[1:50], Daily_interp$Q[1:50],
     col = as.factor(Daily_interp$Qualifier[1:50]),
     main = "Linear Interpolation",
     type = "b", pch = 16, ylim = c(0, 3.2))

Daily_log_interp <- populateDaily(dataInput,
                              qConvert = 1,
                              fill = TRUE,
                              fill_type = "log_interp")
plot(Daily_log_interp$Date, Daily_log_interp$Q,
     col = as.factor(Daily_log_interp$Qualifier),
     main = "Linear Interpolation in Log Scale",
     type = "b", pch = 16, ylim = c(0, 3.2))
plot(Daily_log_interp$Date[1:50], Daily_log_interp$Q[1:50],
     col = as.factor(Daily_log_interp$Qualifier[1:50]),
     main = "Linear Interpolation in Log Scale",
     type = "b", pch = 16, ylim = c(0, 3.2))


# Real data:
eList <- Choptank_eList
Daily_chop <- eList$Daily
df <- Daily_chop[,c("Date", "Q")]
df <- df[-2:-5, ]
df <- df[-100:-200,]
D2 <- populateDaily(df, 1, fill = TRUE)
plot(D2$Date[1:20], D2$Q[1:20],
     col = as.factor(D2$Qualifier[1:20]),
     main = "Linear Interpolation",
     type = "b", pch = 16)
plot(D2$Date[1:110], D2$Q[1:110],
     col = as.factor(D2$Qualifier[1:110]),
     main = "Linear Interpolation",
     type = "b", pch = 16)

}
\seealso{
\code{\link{readNWISDaily}}, \code{\link{readUserDaily}}
}
\author{
Robert M. Hirsch \email{rhirsch@usgs.gov}
}
\keyword{WRTDS}
\keyword{flow}
