% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_functions.R
\name{populateDailySamp}
\alias{populateDailySamp}
\title{Merge concentration to Daily}
\usage{
populateDailySamp(eList)
}
\arguments{
\item{eList}{named list with the INFO, Daily, and Sample dataframes
and surfaces matrix}
}
\description{
Used for the WRTDS Kalman set of functions, this function
merges the ConcAve into the Daily data frame, renaming
it "trueConc", then calculates the "trueFlux", and "stdResid".
}
\examples{
eList <- Choptank_eList
Daily2 <- populateDailySamp(eList)

}
