% Generated by roxygen2 (4.0.1): do not edit by hand
\name{printSeries}
\alias{printSeries}
\title{Print annual results for a given streamflow statistic}
\usage{
printSeries(istat, qUnit = 1, runoff = FALSE, localINFO = INFO,
  localAnnualSeries = annualSeries)
}
\arguments{
\item{istat}{A numeric value for the flow statistic to be graphed (possible values are 1 through 8)}

\item{qUnit}{object of qUnit class \code{\link{qConst}}, or numeric represented the short code, or character representing the descriptive name.}

\item{runoff}{logical variable, if TRUE the streamflow data are converted to runoff values in mm/day}

\item{localINFO}{data frame that contains the metadata, default is INFO}

\item{localAnnualSeries}{data frame containing the annual series, default is AnnualSeries}
}
\description{
Part of the flowHistory system.
The index of the flow statistics is istat.  These statistics are:
(1) 1-day minimum, (2) 7-day minimum, (3) 30-day minimum, (4) median
(5) mean, (6) 30-day maximum, (7) 7-day maximum, and (8) 1-day maximum.
 User must have run setPA and makeAnnualSeries before this function.
}
\examples{
INFO <- ChopINFO
Daily <- ChopDaily
annualSeries <- makeAnnualSeries()
printSeries(5)
}
\keyword{statistics}
\keyword{streamflow}

