% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFFN.R
\name{runFFN}
\alias{runFFN}
\title{runFFN}
\usage{
runFFN(eList, windowSide = NA, dateInfo = NA, verbose = TRUE,
  interactive = FALSE)
}
\arguments{
\item{eList}{named list with at least Daily and INFO dataframes}

\item{windowSide}{integer number of automatically generated span sections, 
default is 7. If NA, code will use}

\item{dateInfo}{data frame with 4 columns. The column names and descriptions are described in the next set of arguments}

\item{verbose}{logical specifying whether or not to display progress message}

\item{interactive}{logical, defaults to FALSE. If TRUE, walks user through options.}
}
\description{
runFFN
}
\examples{
eList <- Choptank_Phos

flowNormStart <- c("1979-10-01","1990-01-01","1992-10-10")
flowNormEnd <- c("1995-06-06","2004-03-03","2011-09-29")
flowStart <- c("1979-10-01","1995-06-07","2004-03-04")
flowEnd <- c("1995-06-06","2004-03-03","2011-09-29") 

dateInfo <- data.frame(flowNormStart,
                       flowNormEnd,
                       flowStart, 
                       flowEnd, 
                       stringsAsFactors = FALSE)
\dontrun{
AnnualResultsFlex <- runFFN(eList, dateInfo=dateInfo)

AnnualResultsFlex <- runFFN(eList, windowSide = 7)

AnnualResultsFlex <- runFFN(eList, interactive = TRUE)

eListWinter <- setPA(eList, paStart = 12, paLong=3)
AnnualResultsFlexWinter <- runFFN(eList, dateInfo)
}
}
