% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPairs.R
\name{runPairs}
\alias{runPairs}
\title{runPairs}
\usage{
runPairs(eList, year1, year2, minNumObs = 100, minNumUncen = 50,
  windowSide = 7, wall = FALSE, lastDaySample1 = NA,
  firstDaySample2 = NA, lastDaySample2 = NA, firstQDate1 = NA,
  lastQDate1 = NA, firstQDate2 = NA, lastQDate2 = NA, ...)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{year1}{integer year1}

\item{year2}{integer year2}

\item{minNumObs}{numeric specifying the miniumum number of observations required to run the weighted regression, default is 100}

\item{minNumUncen}{numeric specifying the minimum number of uncensored observations to run the weighted regression, default is 50}

\item{\dots}{additional parameters}
}
\description{
runPairs description
}
\examples{
eList <- Choptank_eList
year1 <- 1985
year2 <- 2010

pairOut <- runPairs(eList, year1, year2)
}
