% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPairs.R
\name{runPairs}
\alias{runPairs}
\title{runPairs}
\usage{
runPairs(eList, year1, year2, windowSide, flowBreak = FALSE, Q1EndDate = NA,
  QStartDate = NA, QEndDate = NA, wall = FALSE, sample1EndDate = NA,
  sampleStartDate = NA, sampleEndDate = NA, paStart = 10, paLong = 12,
  minNumObs = 100, minNumUncen = 50, windowY = 7, windowQ = 2,
  windowS = 0.5, edgeAdjust = TRUE)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{year1}{integer the ending year of the first year in pairs}

\item{year2}{integer the ending year of the second year in pairs}

\item{windowSide}{integer The width of the flow normalization window on each side of the year being estimated.}

\item{flowBreak}{logical, is there an abrupt break in the QD}

\item{Q1EndDate}{The Date (or character in YYYY-MM-DD) just before the flowBreak (or character in YYY-MM-DD format)}

\item{QStartDate}{The first Date (or character in YYYY-MM-DD) used in the QD (if NA, which is default, it is first Date in eList$Daily)}

\item{QEndDate}{The last Date (or character in YYYY-MM-DD) used in the QD (if NA, which is default, it is the last Date in eList$Daily)}

\item{wall}{logical, there is an abrupt break in the CQR}

\item{sample1EndDate}{The Date (or character in YYYY-MM-DD) of just before the wall}

\item{sampleStartDate}{The Date (or character in YYYY-MM-DD) of the first sample to be used (if NA, which is default, it is the first Date in eList$Sample)}

\item{sampleEndDate}{The Date (or character in YYYY-MM-DD) of the last sample to be used (if NA, which is default, it is the last Date in eList$Sample)}

\item{paStart}{numeric integer specifying the starting month for the period of analysis, 1<=paStart<=12, default is 10}

\item{paLong}{numeric integer specifying the length of the period of analysis, in months, 1<=paLong<=12, default is 12}

\item{minNumObs}{numeric specifying the miniumum number of observations required to run the weighted regression, default is 100}

\item{minNumUncen}{numeric specifying the minimum number of uncensored observations to run the weighted regression, default is 50}

\item{windowY}{numeric specifying the half-window width in the time dimension, in units of years, default is 7}

\item{windowQ}{numeric specifying the half-window width in the discharge dimension, units are natural log units, default is 2}

\item{windowS}{numeric specifying the half-window with in the seasonal dimension, in units of years, default is 0.5}

\item{edgeAdjust}{logical specifying whether to use the modified method for calculating the windows at the edge of the record.
The modified method tends to reduce curvature near the start and end of record.  Default is TRUE.}
}
\description{
runPairs description
}
\examples{
eList <- Choptank_Phos
year1 <- 1985
year2 <- 2014

\dontrun{
# Automatic calculations based on windowSide=7

#Option 1:
pairOut_1 <- runPairs(eList, year1, year2, windowSide = 0)

# Option 2:
pairOut_2 <- runPairs(eList, year1, year2, windowSide = 7)

# Option 3:
pairOut_3 <- runPairs(eList, year1, year2, 
                      windowSide = 0, flowBreak = TRUE,
                      Q1EndDate = "1990-09-30")

# Option 4:
pairOut_4 <- runPairs(eList, year1, year2, 
                      windowSide = 7, flowBreak = TRUE,
                      Q1EndDate = "1990-09-30")

}
}
