% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPairs.R
\name{runPairs}
\alias{runPairs}
\title{runPairs}
\usage{
runPairs(eList, year1, year2, windowSide = 7, wall = FALSE,
  lastDaySample1 = NA, firstQDate0 = NA, lastQDate0 = NA,
  firstQDate1 = NA, lastQDate1 = NA, firstQDate2 = NA, lastQDate2 = NA,
  minNumObs = 100, minNumUncen = 50, windowY = 7, windowQ = 2,
  windowS = 0.5, interactive = FALSE, edgeAdjust = TRUE, ...)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{year1}{integer year1}

\item{year2}{integer year2}

\item{windowSide}{integer number of automatically generated span sections, 
default is 7. If NA, code will use}

\item{wall}{logical set up a "wall" on the Sample data}

\item{lastDaySample1}{character in YYYY-MM-DD. Only used if wall is TRUE. Date of "the wall".}

\item{firstQDate0}{character in YYYY-MM-DD. Overall trims Daily flow. Use NA to use all data.}

\item{lastQDate0}{character in YYYY-MM-DD. Overall trims Daily flow. Use NA to use all data.}

\item{firstQDate1}{character in YYYY-MM-DD. First limit of flow data to use in year 1. Use NA to automatically calculate based on windowSide.}

\item{lastQDate1}{character in YYYY-MM-DD. Second limit of flow data to use in year 1. Use NA to automatically calculate based on windowSide.}

\item{firstQDate2}{character in YYYY-MM-DD. First limit of flow data to use in year 2. Use NA to automatically calculate based on windowSide.}

\item{lastQDate2}{character in YYYY-MM-DD. Second limit of flow data to use in year 2. Use NA to automatically calculate based on windowSide.}

\item{minNumObs}{numeric specifying the miniumum number of observations required to run the weighted regression, default is 100}

\item{minNumUncen}{numeric specifying the minimum number of uncensored observations to run the weighted regression, default is 50}

\item{windowY}{numeric specifying the half-window width in the time dimension, in units of years, default is 7}

\item{windowQ}{numeric specifying the half-window width in the discharge dimension, units are natural log units, default is 2}

\item{windowS}{numeric specifying the half-window with in the seasonal dimension, in units of years, default is 0.5}

\item{interactive}{logical, defaults to FALSE. If TRUE, walks user through options.}

\item{edgeAdjust}{logical specifying whether to use the modified method for calculating the windows at the edge of the record.
The modified method tends to reduce curvature near the start and end of record.  Default is TRUE.}

\item{\dots}{additional parameters}

\item{minNumObs}{numeric specifying the miniumum number of observations required to run the weighted regression, default is 100}

\item{minNumUncen}{numeric specifying the minimum number of uncensored observations to run the weighted regression, default is 50}
}
\description{
runPairs description
}
\examples{
eList <- Choptank_Phos
year1 <- 1985
year2 <- 2014

\dontrun{
# Automatic calculations based on windowSide=7
pairOut <- runPairs(eList, year1, year2)

# Specify flow normalize sections:
pairOut_custom <- runPairs(eList, year1, year2,
                           firstQDate1 = "1977-10-01",
                           lastQDate1 = "1990-09-30",
                           firstQDate2 = "2000-10-01",
                           lastQDate2 = "2014-09-25") 

pairOut_interactive <- runPairs(eList, year1, year2,
                                interactive = TRUE)

}
}
