% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSeries.R
\name{runSeries}
\alias{runSeries}
\title{runSeries}
\usage{
runSeries(eList, surfaceStart = NA, surfaceEnd = NA, wall = FALSE,
  firstSampleDate0 = NA, lastSampleDate0 = NA, lastSampleDate1 = NA,
  flowBreak = FALSE, windowSide = 7, lastQDate1 = NA, firstQDate0 = NA,
  lastQDate0 = NA, minNumObs = 100, minNumUncen = 50, windowY = 7,
  windowQ = 2, windowS = 0.5, edgeAdjust = TRUE)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{surfaceStart}{character (or Date) in YYYY-MM-DD. Date on which we want the analysis to start, it must be at or after the}

\item{surfaceEnd}{character (or Date) in YYYY-MM-DD. Date on which we want the analysis to end, it must be at or before the end of}

\item{wall}{logical set up a "wall" on the Sample data}

\item{firstSampleDate0}{character (or Date) in YYYY-MM-DD. First day of the Sample data set we will use, default is 
use all samples starting with the first sample}

\item{lastSampleDate0}{character (or Date) in YYYY-MM-DD. Last day of the Sample data set we will use, default is 
use up through the last sample}

\item{lastSampleDate1}{character (or Date) in YYYY-MM-DD. Day just before the wall in the RS}

\item{flowBreak}{logical. User-specified break in the flow distribution}

\item{windowSide}{integer number of automatically generated span sections, 
default is 7. If NA, code will use}

\item{lastQDate1}{character (or Date) in YYYY-MM-DD. The last day of the period leading up to the flow break, must be specified if
flowBreak = TRUE}

\item{firstQDate0}{character (or Date) in YYYY-MM-DD. The first day used in flow normalizing distributions, default is 
the start of eList$Daily}

\item{lastQDate0}{character (or Date) in YYYY-MM-DD. The last day used in flow normalizating distributions, default is
the end of eList$Daily}
}
\description{
runSeries
four possible situations 
 1 no flexible flow normalization at all
 2 moving window flexible flow normalization, but no break
 3 there is a break (must be specified as lastQDate1) but no moving window on each side
 4 there is a break (must be specified as lastQDate1) but there is a moving window on each side
}
\examples{
eList <- Choptank_Phos


}
