% Generated by roxygen2 (4.0.2): do not edit by hand
\name{saveResults}
\alias{saveResults}
\title{A utility program for saving the contents of the workspace}
\usage{
saveResults(savePath, eList)
}
\arguments{
\item{savePath}{character specifying the full pathname of the folder where the file is to be saved ending with the final slash}

\item{eList}{named list with at least the INFO dataframe}
}
\description{
This function saves the workspace.
 It assigns the file a name using the abbreviations for station and constituent.
}
\examples{
eList <- Choptank_eList
savePath <- "~/"
saveResults(savePath, eList)

#To load:
#load(paste(savePath,"Chop.nitrogen.RData",sep=""))
}
\keyword{statistics}
\keyword{water-quality}

