% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch}
\alias{stitch}
\title{stitch}
\usage{
stitch(eList, firstSampleDate1, lastSampleDate1, firstSampleDate2,
  lastSampleDate2, surfaceStart = NA, surfaceEnd = NA, minNumObs = 100,
  minNumUncen = 50, windowY = 7, windowQ = 2, windowS = 0.5,
  edgeAdjust = TRUE, verbose = FALSE, run.parallel = FALSE)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{firstSampleDate1}{character in YYYY-MM-DD. Starting data of Sample data for first surface to be joined.}

\item{lastSampleDate1}{character in YYYY-MM-DD. Ending data of Sample data for first surface to be joined.}

\item{firstSampleDate2}{character in YYYY-MM-DD. Starting data of Sample data for second surface to be joined.}

\item{lastSampleDate2}{character in YYYY-MM-DD. Ending data of Sample data for second surface to be joined.}

\item{surfaceStart}{character (or Date) in YYYY-MM-DD. Starting date of final surface.
If NA, surface starts at beginning of Sample}

\item{surfaceEnd}{character (or Date) in YYYY-MM-DD. Ending date of final surface
If NA, surface ends at end of Sample}

\item{minNumObs}{numeric specifying the miniumum number of observations required to run the weighted regression, default is 100}

\item{minNumUncen}{numeric specifying the minimum number of uncensored observations to run the weighted regression, default is 50}

\item{windowY}{numeric specifying the half-window width in the time dimension, in units of years, default is 7}

\item{windowQ}{numeric specifying the half-window width in the discharge dimension, units are natural log units, default is 2}

\item{windowS}{numeric specifying the half-window with in the seasonal dimension, in units of years, default is 0.5}

\item{edgeAdjust}{logical specifying whether to use the modified method for calculating the windows at the edge of the record.
The modified method tends to reduce curvature near the start and end of record.  Default is TRUE.}

\item{verbose}{logical specifying whether or not to display progress message}

\item{run.parallel}{logical to run bootstrapping in parallel or not}
}
\description{
This function creates a continuous surfaces object that starts just before 
surfaceStart and ends just after surfaceEnd.
}
\details{
It has a wall in it and the wall comes just after lastDaySample1
Arguments 2 through 5 are all dates, but just to make sure we will do a conversion to as.Date
the only use being made of eList$Daily is to establish the minimum and maximum discharges for the surfaces
the eList$Sample will be split out into Sample1 and Sample2 (there will be no overlap)
}
\examples{
eList <- Choptank_Phos

surfaceStart <- "1986-10-01"
surfaceEnd <- "2012-09-30"

# Surface skips a few years:
firstSampleDate1 <- "1986-10-01"
lastSampleDate1 <- "1992-09-30"
firstSampleDate2 <- "1996-10-01"
lastSampleDate2 <- "2012-09-30"

\dontrun{
surface_skip <- stitch(eList, 
                         surfaceStart, surfaceEnd, 
                         firstSampleDate1, lastSampleDate1, 
                         firstSampleDate2, lastSampleDate2)
   
eList_skip <- as.egret(eList$INFO, 
                       eList$Daily, 
                       eList$Sample, 
                       surface_skip)
eList_skip <- modelEstimation(eList_skip)
plotConcHist(eList_skip)

# Surface overlaps a few years:
firstSampleDate1 <- "1986-10-01"
lastSampleDate1 <- "1996-09-30"
firstSampleDate2 <- "1992-10-01"
lastSampleDate2 <- "2012-09-30"

surface_overlap <- stitch(eList, 
                         surfaceStart, surfaceEnd, 
                         firstSampleDate1, lastSampleDate1, 
                         firstSampleDate2, lastSampleDate2)
}
}
