% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/surfaceIndex.R
\name{surfaceIndex}
\alias{surfaceIndex}
\title{Compute the 6 parameters needed to lay out the grid for the surfaces computed in estSurfaces}
\usage{
surfaceIndex(Daily)
}
\arguments{
\item{Daily}{data frame containing the daily values, default is Daily}
}
\value{
surfaceIndexParameters a numeric vector of length 6, defining the grid for the surfaces
}
\description{
The code here is a repetition of the first part of the code for estSurfaces
}
\examples{
eList <- Choptank_eList
Daily <- getDaily(eList)
surfaceIndex(Daily)
}
\keyword{statistics}
\keyword{water-quality}

