% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_functions.R
\name{WRTDSKalman}
\alias{WRTDSKalman}
\title{WRTDS-Kalman}
\usage{
WRTDSKalman(eList, rho = 0.9, niter = 200, seed = 376168)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{rho}{numeric the lag one autocorrelation. Default is 0.9.}

\item{niter}{number of iterations. Default is 200.}

\item{seed}{setSeed value. Defaults to 376168 This is used to make repeatable output.}
}
\description{
This function takes an existing eList 
Including the estimated model (the surfaces object in the eList)
And produces the daily WRTDSKalman estimates of concentration and flux
}
\examples{
eList <- Choptank_eList
eList <- WRTDSKalman(eList)
summary(eList$Daily)

AnnualResults <- setupYears(eList$Daily)
head(AnnualResults)
}
