% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_functions.R
\name{WRTDS_K}
\alias{WRTDS_K}
\alias{random_subset}
\alias{plotWRTDS_K}
\alias{plotTimeSlice}
\title{WRTDS-K}
\usage{
WRTDS_K(eList, rho = 0.9, niter = 200, seed = 376168)

random_subset(df, col_name)

plotWRTDS_K(eList)

plotTimeSlice(eList, start, end, conc = TRUE, fluxUnit = 3, usgsStyle = FALSE)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{rho}{numeric}

\item{niter}{numeric}

\item{seed}{number to set seed for reproducibility}

\item{df}{data frame}

\item{col_name}{column name}

\item{start}{numeric start of DecYear for plot}

\item{end}{numeric end of DecYear for plot}

\item{conc}{logical. If \code{TRUE}, plot concentration, otherwise plot flux.}

\item{fluxUnit}{number representing entry in pre-defined fluxUnit class array. \code{\link{printFluxUnitCheatSheet}}}

\item{usgsStyle}{logical option to use USGS style guidelines. Setting this option
to TRUE does NOT guarantee USGS compliance. It will only change automatically
generated labels}
}
\description{
This function takes an existing eList 
Including the estimated model (the surfaces object in the eList)
And produces the daily WRTDS_K estimates of concentration and flux
}
\examples{
eList <- Choptank_eList
eList <- WRTDS_K(eList)
summary(eList$Daily)

AnnualResults <- setupYears(eList$Daily)
head(AnnualResults)
df <- data.frame(Julian = c(1,2,2,3,4,4,4,6),
                 y = 1:8)
df
df_random <- random_subset(df, "Julian")
df_random
plotWRTDS_K(eList)
plotTimeSlice(eList, start = 1990, end = 1991, conc = TRUE)

plotTimeSlice(eList, start = 1990, end = 1991, conc = FALSE)

}
