% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_functions.R
\name{makeDailyK}
\alias{makeDailyK}
\alias{cleanUp}
\alias{random_subset}
\alias{populateDailySamp}
\alias{genmissing}
\alias{computeAnnual}
\alias{plotTimeSlice}
\title{WRTDS-K}
\usage{
makeDailyK(eList, rho = 0.9, niter = 200, seed = 376168)

cleanUp(eList)

random_subset(df, col_name)

populateDailySamp(eList)

genmissing(X1, XN, rho, N)

computeAnnual(eList, DailyK, paStart = 10, paLong = 12)

plotTimeSlice(eList, DailyK, start, end)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{rho}{the lag one autocorrelation}

\item{niter}{numeric}

\item{seed}{number to set seed for reproducibility}

\item{df}{data frame}

\item{col_name}{column name}

\item{X1}{the value before the gap}

\item{XN}{value after the gap}

\item{N}{the length of the sequence including X1 and XN 
it is two more than the gap length}

\item{DailyK}{Daily data frame run through the WRTDS-K analysis}

\item{paStart}{A numeric value for the starting month of the Period of Analysis, default is 10}

\item{paLong}{A numeric value for the length of the Period of Analysis in months, default is 12}

\item{start}{numeric start of DecYear for plot}

\item{end}{numeric end of DecYear for plot}
}
\description{
This function takes an existing eList 
Including the estimated model (the surfaces object in the eList)
And produces the daily WRTDS_K estimates of concentration and flux
}
\examples{
eList <- Choptank_eList
DailyK <- makeDailyK(eList)
summary(DailyK)

AnnualResults <- setupYears(DailyK)


eList <- cleanUp(eList)

eList <- Choptank_eList
Sample <- eList$Sample
random_subset(Sample, col_name = "Julian")

df <- data.frame(Julian = c(1,2,2,3,4,4,4,6),
                 y = 1:8)
random_subset(df, "Julian")

d2 <- populateDailySamp(eList)




computeAnnual(eList, DailyK)
plotTimeSlice(eList, DailyK, start = 1990, end = 1991)
}
