% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_functions.R
\name{WRTDSKalman}
\alias{WRTDSKalman}
\title{WRTDS-Kalman}
\usage{
WRTDSKalman(eList, rho = 0.9, niter = 200, seed = NA, verbose = TRUE)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes and surfaces matrix}

\item{rho}{numeric the lag one-day autocorrelation. Default is 0.9.}

\item{niter}{number of iterations. Default is 200.}

\item{seed}{setSeed value. Defaults to NA, which will not change the current seed.
This is used to make repeatable output.}

\item{verbose}{logical specifying whether or not to display progress message}
}
\description{
This function takes an existing eList 
Including the estimated model (the surfaces object in the eList)
And produces the daily WRTDSKalman estimates of concentration and flux
The function returns an eList in which the estimates are appended to the existing Daily dataframe
}
\examples{
eList <- Choptank_eList
eList <- WRTDSKalman(eList, niter = 10)
summary(eList$Daily)

AnnualResults <- setupYears(eList$Daily)
head(AnnualResults)
}
